Support cross-entropy loss in grad step.
Initially tried taking a std::function for the loss gradient, but this
caused a slowdown of 20% or so.

This is likely due to the overhead of using Python function callbacks as
std::function arguments via pybind11. The overhead would probably be
minimal in pure C++ code.

A more flexible solution would be to take std::function (or even a
template function type) at C++ level, but provide some standard losses
in the package, so the Python caller can refer to our losses instead of
going through the Python callback interface.